Added hand gestures
Added the hand gesture python, but now working yet.
Also made navigation nice with animations.

For next:
Save user globally with a face detection to check if there is still a user and if it needs to be logged out.
Make handgesture even smoother
Add handgesture to main script, not rendered scripts.